secp256k1_omni_context: Rename
Previous name was too annoying to type so went for a shorter name
