+++ internal/bitcoin_witness/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2018"
 [dependencies]
 bitcoin_support = { path = "../bitcoin_support" }
 failure = "0.1"
-secp256k1_omni_context = { path = "../secp256k1_omni_context" }
+secp_wrapper = { path = "../secp_wrapper" }
 
 [dev-dependencies]
 bitcoin_rpc_test_helpers = { path = "../bitcoin_rpc_test_helpers" }
