+++ internal/bitcoin_witness/src/lib.rs
@@ -19,4 +19,4 @@ pub use crate::{
     witness::{UnlockParameters, Witness},
 };
 pub use secp256k1::PublicKey;
-pub use secp256k1_omni_context::{secp256k1, Builder, SecretKey};
+pub use secp_wrapper::{secp256k1, Builder, SecretKey};
