+++ internal/key_gen/Cargo.toml
@@ -9,5 +9,5 @@ bitcoin = "0.19.1"
 bitcoin_support  = { path = "../bitcoin_support" }
 ethereum_support = { path = "../ethereum_support" }
 hex = "0.4"
-secp256k1_omni_context = { path = "../secp256k1_omni_context" }
+secp_wrapper = { path = "../secp_wrapper" }
 tiny-keccak = "1.5.0"
